+++ webgear-core/src/WebGear/Core/Trait/Auth/Common.hs
@@ -3,13 +3,11 @@
 -}
 module WebGear.Core.Trait.Auth.Common (
   AuthorizationHeader,
-  getAuthorizationHeaderTrait,
   Realm (..),
   AuthToken (..),
   respondUnauthorized,
 ) where
 
-import Control.Arrow (returnA)
 import Data.ByteString (ByteString, drop)
 import Data.ByteString.Char8 (break)
 import Data.CaseInsensitive (CI, mk, original)
@@ -17,15 +15,13 @@ import Data.Proxy (Proxy (..))
 import Data.String (IsString (..))
 import Data.Text (Text)
 import Data.Text.Encoding (decodeUtf8, encodeUtf8)
-import Data.Void (absurd)
 import GHC.TypeLits (KnownSymbol, Symbol, symbolVal)
 import Web.HttpApiData (FromHttpApiData (..))
 import WebGear.Core.Handler (Handler, unwitnessA, (>->))
 import WebGear.Core.MIMETypes (PlainText (..))
 import WebGear.Core.Modifiers (Existence (..), ParseStyle (..))
-import WebGear.Core.Request (Request)
 import WebGear.Core.Response (Response)
-import WebGear.Core.Trait (Get (..), Sets, With)
+import WebGear.Core.Trait (Sets)
 import WebGear.Core.Trait.Body (Body, setBody)
 import WebGear.Core.Trait.Header (RequestHeader (..), RequiredResponseHeader, setHeader)
 import WebGear.Core.Trait.Status (Status, unauthorized401)
@@ -33,20 +29,6 @@ import Prelude hiding (break, drop)
 
 -- | Trait for \"Authorization\" header
 type AuthorizationHeader scheme = RequestHeader Optional Lenient "Authorization" (AuthToken scheme)
-
-{- | Extract the \"Authorization\" header from a request by specifying
-   an authentication scheme.
-
-  The header is split into the scheme and token parts and returned.
--}
-getAuthorizationHeaderTrait ::
-  forall scheme h ts.
-  (Get h (AuthorizationHeader scheme) Request) =>
-  h (Request `With` ts) (Maybe (Either Text (AuthToken scheme)))
-getAuthorizationHeaderTrait = proc request -> do
-  result <- getTrait (RequestHeader :: RequestHeader Optional Lenient "Authorization" (AuthToken scheme)) -< request
-  returnA -< either absurd id result
-{-# INLINE getAuthorizationHeaderTrait #-}
 
 -- | The protection space for authentication
 newtype Realm = Realm ByteString
@@ -61,9 +43,12 @@ data AuthToken (scheme :: Symbol) = AuthToken
   }
 
 instance (KnownSymbol scheme) => FromHttpApiData (AuthToken scheme) where
+  {-# INLINE parseUrlPiece #-}
+  parseUrlPiece :: Text -> Either Text (AuthToken scheme)
   parseUrlPiece = parseHeader . encodeUtf8
 
   {-# INLINE parseHeader #-}
+  parseHeader :: ByteString -> Either Text (AuthToken scheme)
   parseHeader hdr =
     case break (== ' ') hdr of
       (scm, tok) ->